+++ src/GitHub_Updater/API/API_Interface.php
@@ -8,7 +8,7 @@
  * @link      https://github.com/afragen/github-updater
  */
 
-namespace Fragen\GitHub_Updater;
+namespace Fragen\GitHub_Updater\API;
 
 
 /**