+++ src/GitHub_Updater/Branch.php
@@ -10,6 +10,9 @@
 
 namespace Fragen\GitHub_Updater;
 
+use Fragen\Singleton;
+
+
 /*
  * Exit if called directly.
  */
@@ -48,7 +51,7 @@ class Branch {
 	 */
 	public function __construct( $cache = null ) {
 		$this->cache     = $cache;
-		$base            = \Fragen\Singleton::get_instance( 'Base' );
+		$base            = Singleton::get_instance( 'Base' );
 		static::$options = $base::$options;
 	}
 
@@ -77,7 +80,7 @@ class Branch {
 	 * @param string $repo Repository slug.
 	 */
 	public function set_branch_on_switch( $repo ) {
-		$this->cache = \Fragen\Singleton::get_instance( 'API_PseudoTrait' )->get_repo_cache( $repo );
+		$this->cache = Singleton::get_instance( 'API_PseudoTrait' )->get_repo_cache( $repo );
 
 		if ( isset( $_GET['action'], $_GET['rollback'], $this->cache['branches'] ) &&
 		     ( 'upgrade-plugin' === $_GET['action'] || 'upgrade-theme' === $_GET['action'] )
@@ -85,7 +88,7 @@ class Branch {
 			$current_branch = array_key_exists( $_GET['rollback'], $this->cache['branches'] )
 				? $_GET['rollback']
 				: 'master';
-			\Fragen\Singleton::get_instance( 'API_PseudoTrait' )->set_repo_cache( 'current_branch', $current_branch, $repo );
+			Singleton::get_instance( 'API_PseudoTrait' )->set_repo_cache( 'current_branch', $current_branch, $repo );
 			static::$options[ 'current_branch_' . $repo ] = $current_branch;
 			update_site_option( 'github_updater', static::$options );
 		}
@@ -99,7 +102,7 @@ class Branch {
 	 * @param array $install Array of install data.
 	 */
 	public function set_branch_on_install( $install ) {
-		\Fragen\Singleton::get_instance( 'API_PseudoTrait' )->set_repo_cache( 'current_branch', $install['github_updater_branch'], $install['repo'] );
+		Singleton::get_instance( 'API_PseudoTrait' )->set_repo_cache( 'current_branch', $install['github_updater_branch'], $install['repo'] );
 		static::$options[ 'current_branch_' . $install['repo'] ] = $install['github_updater_branch'];
 		update_site_option( 'github_updater', static::$options );
 	}