+++ src/GitHub_Updater/Language_Pack.php
@@ -10,6 +10,10 @@
 
 namespace Fragen\GitHub_Updater;
 
+use Fragen\Singleton,
+	Fragen\GitHub_Updater\API\Language_Pack_API;
+
+
 /**
  * Exit if called directly.
  */
@@ -79,11 +83,11 @@ class Language_Pack extends Base {
 		}
 
 		if ( 'pre_set_site_transient_update_plugins' === current_filter() ) {
-			$repos        = \Fragen\Singleton::get_instance( 'Plugin' )->get_plugin_configs();
+			$repos        = Singleton::get_instance( 'Plugin' )->get_plugin_configs();
 			$translations = wp_get_installed_translations( 'plugins' );
 		}
 		if ( 'pre_set_site_transient_update_themes' === current_filter() ) {
-			$repos        = \Fragen\Singleton::get_instance( 'Theme' )->get_theme_configs();
+			$repos        = Singleton::get_instance( 'Theme' )->get_theme_configs();
 			$translations = wp_get_installed_translations( 'themes' );
 		}