+++ src/GitHub_Updater/Readme_Parser.php
@@ -11,8 +11,9 @@
 
 namespace Fragen\GitHub_Updater;
 
-use WordPressdotorg\Plugin_Directory\Readme\Parser;
-use Parsedown;
+use WordPressdotorg\Plugin_Directory\Readme\Parser,
+	Parsedown;
+
 
 /*
  * Exit if called directly.