+++ src/GitHub_Updater/Rest_Upgrader_Skin.php
@@ -11,6 +11,7 @@
 
 namespace Fragen\GitHub_Updater;
 
+
 /*
  * Exit if called directly.
  */