+++ src/GitHub_Updater/WP-CLI/CLI_Integration.php
@@ -11,7 +11,9 @@
 namespace Fragen\GitHub_Updater;
 
 use WP_CLI,
-	WP_CLI_Command;
+	WP_CLI_Command,
+	Fragen\Singleton;
+
 
 // Add WP-CLI commands.
 $class = new CLI_Integration();
@@ -34,7 +36,7 @@ class CLI_Integration extends WP_CLI_Command {
 	 * GitHub_Updater_CLI_Integration constructor.
 	 */
 	public function __construct() {
-		$this->base = \Fragen\Singleton::get_instance( 'Base' );
+		$this->base = Singleton::get_instance( 'Base' );
 		$this->run();
 	}
 
@@ -55,7 +57,7 @@ class CLI_Integration extends WP_CLI_Command {
 	public function init_plugins() {
 		$this->base->get_meta_plugins();
 		$current = get_site_transient( 'update_plugins' );
-		$current = \Fragen\Singleton::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
+		$current = Singleton::get_instance( 'Plugin' )->pre_set_site_transient_update_plugins( $current );
 		set_site_transient( 'update_plugins', $current );
 	}
 
@@ -68,7 +70,7 @@ class CLI_Integration extends WP_CLI_Command {
 	public function init_themes() {
 		$this->base->get_meta_themes();
 		$current = get_site_transient( 'update_themes' );
-		$current = \Fragen\Singleton::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
+		$current = Singleton::get_instance( 'Theme' )->pre_set_site_transient_update_themes( $current );
 		set_site_transient( 'update_themes', $current );
 	}
 
@@ -122,7 +124,7 @@ class CLI_Integration extends WP_CLI_Command {
 	public function install_plugin( $args, $assoc_args ) {
 		list( $uri ) = $args;
 		$cli_config = $this->process_args( $uri, $assoc_args );
-		\Fragen\Singleton::get_instance( 'Install' )->install( 'plugin', $cli_config );
+		Singleton::get_instance( 'Install' )->install( 'plugin', $cli_config );
 
 		$headers = parse_url( $uri, PHP_URL_PATH );
 		$slug    = basename( $headers );
@@ -180,7 +182,7 @@ class CLI_Integration extends WP_CLI_Command {
 	public function install_theme( $args, $assoc_args ) {
 		list( $uri ) = $args;
 		$cli_config = $this->process_args( $uri, $assoc_args );
-		\Fragen\Singleton::get_instance( 'Install' )->install( 'theme', $cli_config );
+		Singleton::get_instance( 'Install' )->install( 'theme', $cli_config );
 
 		$headers = parse_url( $uri, PHP_URL_PATH );
 		$slug    = basename( $headers );
@@ -231,7 +233,7 @@ class CLI_Integration extends WP_CLI_Command {
 		$branch_data['github_updater_branch'] = $cli_config['branch'];
 		$branch_data['repo']                  = $slug;
 
-		\Fragen\Singleton::get_instance( 'Branch' )->set_branch_on_install( $branch_data );
+		Singleton::get_instance( 'Branch' )->set_branch_on_install( $branch_data );
 	}
 
 }