+++ src/Singleton.php
@@ -38,13 +38,7 @@ if ( ! class_exists( 'Fragen\\Singleton' ) ) {
 			static $instance = null;
 			$backtrace = debug_backtrace();
 			$class     = isset( $backtrace[1]['class'] ) ? $backtrace[1]['class'] : null;
-
-			try {
-				$reflection = new \ReflectionClass( $class );
-				$class      = $reflection->getNamespaceName() . '\\' . $class_name;
-			} catch ( \ReflectionException $Exception ) {
-				die( '<table>' . $Exception->xdebug_message . '</table>' );
-			}
+			$class     = self::get_class( $class_name, $class );
 
 			if ( null === $instance || ! isset( $instance[ $class ] ) ) {
 				$instance[ $class ] = new $class( $options );
@@ -55,6 +49,27 @@ if ( ! class_exists( 'Fragen\\Singleton' ) ) {
 
 			return $instance[ $class ];
 		}
-	}
 
+		/**
+		 * Determine correct class name with namespace and return.
+		 *
+		 * @param string $class_name
+		 * @param string $class
+		 *
+		 * @return string $class
+		 */
+		private static function get_class( $class_name, $class ) {
+			try {
+				$reflection      = new \ReflectionClass( $class );
+				$namespace       = $reflection->getNamespaceName();
+				$namespace_parts = explode( '\\', $namespace );
+				$namespace_root  = $namespace_parts[0] . '\\' . $namespace_parts[1];
+				$class           = $namespace_root . '\\' . $class_name;
+			} catch ( \ReflectionException $Exception ) {
+				die( '<table>' . $Exception->xdebug_message . '</table>' );
+			}
+
+			return $class;
+		}
+	}
 }